move getChangelogVersionData to changelog utils
