+++ src/utils/data.ts
@@ -55,32 +55,6 @@ export function getBlogMetaData({ type, filterValue }: BlogMetadataParams) {
   };
 }
 
-export function getChangelogVersionData(
-  currentVersion: string,
-  allReleases: string[]
-) {
-  const index = allReleases.findIndex(version => version === currentVersion);
-  const newerVersion = index === 0 ? null : allReleases[index - 1];
-  const olderVersion =
-    index === allReleases.length - 1 ? null : allReleases[index + 1];
-
-  const generateChangelogData = (version: string | null) => {
-    if (version == null) {
-      return null;
-    }
-
-    return {
-      slug: `/changelog/${version}/`,
-      title: version
-    };
-  };
-
-  return {
-    newerData: generateChangelogData(newerVersion),
-    olderData: generateChangelogData(olderVersion)
-  };
-}
-
 export function getOtherMetadata(
   title: string,
   description?: string,
