+++ test/utils/changelog.test.ts
@@ -4,6 +4,7 @@
 
 import {
   getAllReleases,
+  getChangelogVersionData,
   getGithubConstants,
   getReleaseContent
 } from "@/utils/changelog";
@@ -103,5 +104,48 @@ describe("Test getReleaseContent function", () => {
     expect(async () => {
       await getReleaseContent("2022.01.02");
     }).rejects.toThrow("Failed to fetch github releases");
+  });
+});
+
+describe("Test getChangelogVersionData", () => {
+  const allReleasesMockData = ["2022.01.04", "2021.10.28", "2021.08.12"];
+
+  test("First entry", () => {
+    const result = getChangelogVersionData("2022.01.04", allReleasesMockData);
+
+    expect(result).toMatchObject({
+      newerData: null,
+      olderData: {
+        slug: `/changelog/2021.10.28/`,
+        title: `2021.10.28`
+      }
+    });
+  });
+
+  test("Last entry", () => {
+    const result = getChangelogVersionData("2021.08.12", allReleasesMockData);
+
+    expect(result).toMatchObject({
+      newerData: {
+        slug: `/changelog/2021.10.28/`,
+        title: `2021.10.28`
+      },
+      olderData: null
+    });
+  });
+
+  test("Middle entry", () => {
+    const result = getChangelogVersionData("2021.10.28", allReleasesMockData);
+
+    expect(result).toMatchObject({
+      newerData: {
+        slug: `/changelog/2022.01.04/`,
+        title: `2022.01.04`
+      },
+      olderData: {
+        slug: `/changelog/2021.08.12/`,
+        title: `2021.08.12`
+      }
+    });
   });
 });
